fix: rename and set `false` by default
