+++ crates/swc_css_modules/tests/with_compat.rs
@@ -33,7 +33,7 @@ fn test_full(input: PathBuf, suffix: Option<&str>) {
             &fm,
             ParserConfig {
                 css_modules: true,
-                allow_nested_selectors: true,
+                legacy_nesting: true,
                 ..Default::default()
             },
             &mut errors,
