+++ crates/swc_css_parser/src/parser/base.rs
@@ -107,69 +107,6 @@ where
             }
         }
     }
-
-    fn try_parse_qualified_rule(&mut self) -> Option<Box<QualifiedRule>> {
-        if !self.config.allow_nested_selectors {
-            return None;
-        }
-
-        let state = self.input.state();
-
-        let ctx = Ctx {
-            is_trying_nested_selector: true,
-            ..self.ctx
-        };
-
-        let span = self.input.cur_span();
-
-        let nested = self.with_ctx(ctx).parse_as::<Box<QualifiedRule>>();
-
-        let mut nested = match nested {
-            Ok(v) => v,
-            Err(_) => {
-                self.input.reset(&state);
-                return None;
-            }
-        };
-
-        match &mut nested.prelude {
-            QualifiedRulePrelude::ListOfComponentValues(_) => {
-                self.input.reset(&state);
-                return None;
-            }
-            QualifiedRulePrelude::SelectorList(s) => {
-                for s in s.children.iter_mut() {
-                    if s.children.iter().any(|s| match s {
-                        ComplexSelectorChildren::CompoundSelector(s) => {
-                            s.nesting_selector.is_some()
-                        }
-                        _ => false,
-                    }) {
-                        continue;
-                    }
-
-                    s.children.insert(
-                        0,
-                        ComplexSelectorChildren::CompoundSelector(CompoundSelector {
-                            span,
-                            nesting_selector: Some(NestingSelector { span }),
-                            type_selector: Default::default(),
-                            subclass_selectors: Default::default(),
-                        }),
-                    );
-                    s.children.insert(
-                        1,
-                        ComplexSelectorChildren::Combinator(Combinator {
-                            span,
-                            value: CombinatorValue::Descendant,
-                        }),
-                    );
-                }
-            }
-        }
-
-        Some(nested)
-    }
 }
 
 impl<I> Parse<QualifiedRule> for Parser<I>
@@ -278,7 +215,7 @@ where
             //
             //      }
             // }
-            if self.config.allow_nested_selectors {
+            if self.config.legacy_nesting {
                 let state = self.input.state();
                 let nested = self.try_parse_qualified_rule();
